Replace Image Source and Size
The previous source of the image was from the AllRecipe website.
The image has been downloaded and stored in the Images directory to
ensure availability.

The previous size of the image was larger than necessary and forced
users to scroll to view any other content. Image size was reduced to
increase page readability.